HTCONDOR-2389 More repatriation of SafeSock and ReliSock methods
